Replace prev GraphsContainer with new and archive as Archived.tsx
