+++ app/containers/MainContainer.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 
-import GraphsContainer from './GraphsContainer';
+// import GraphsContainer from './Archived';
 import Home from '../components/Home';
 import Copyright from '../components/Copyright';
 import Occupied from '../components/Occupied';
-import TempContainer from './TempContainer';
+import GraphsContainer from './GraphsContainer';
 
 const MainContainer = () => (
   <div style={{ flex: '1' }}>
     <Switch>
       <Route exact path="/" component={Home} />
       <Route exact path="/applications" component={Occupied} />
-      <Route exact path="/applications/:app/:service" component={TempContainer} />
+      <Route exact path="/applications/:app/:service" component={GraphsContainer} />
       <Route path="*" render={() => <h1>Not found</h1>} />
     </Switch>
     <Copyright />
